Change links and styling of sponsors and transport
We changed the links of the sponsors and the transport page. The links in the sponsors page were moved from the top to the bottom.The bus logo on the transport page was aligned properly for mobile. The links on this page were modified and an extra link was added.