Change links (cmd/m), typos (eventsFetch.php)
There was an incorrect link in the cmd/m page and I changed it to the correct one. There were also several typos in the eventsFetch.php file and I changed it. I also changed the name of a CSS file from home-css.css to index-css.css to match the patterns found everywhere else in the site.